+++ packages/remix-image/src/types/transformer.ts
@@ -70,6 +70,7 @@ export type Transformer = {
   name: string;
   supportedInputs: Set<MimeType>;
   supportedOutputs: Set<MimeType>;
+  fallbackOutput: MimeType;
   transform: (
     input: {
       url: string;