+++ packages/remix-image-wasm/CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Set `fallbackOutput` field to first loaded format in the order of `MimeType.PNG`, `MimeType.JPEG`, `MimeType.WEBP`, `MimeType.AVIF`
+- Set `supportedInputs` and `supportedOutputs` to handlers that have had `.wasm` files included in environment
+
 ## [0.1.1] - 2022-06-28
 
 ### Fixed